+++ DBSystemComparator-API/Repositories/Interfaces/IPostgreSQLRepository.cs
@@ -7,10 +7,10 @@ namespace DBSystemComparator_API.Repositories.Interfaces
     {
         // CREATE
         Task<int> CreateClientAsync(string firstName, string secondName, string lastName, string email, DateTime dob, string address, string phone, bool isActive);
-        Task<int> CreateRoomAsync(int number, int capacity, int pricePerNight, bool isActive);
+        Task<int> CreateRoomAsync(int number, int capacity, double pricePerNight, bool isActive);
         Task<int> CreateServiceAsync(string name, int price, bool isActive);
         Task<List<int>> CreateClientsAsync(string firstName, string secondName, string lastName, string email, DateTime dob, string address, string phone, bool isActive, int count);
-        Task<List<int>> CreateRoomsAsync(int number, int capacity, int pricePerNight, bool isActive, int count);
+        Task<List<int>> CreateRoomsAsync(int number, int capacity, double pricePerNight, bool isActive, int count);
         // READ
         Task<List<Dictionary<string, object>>> ReadReservationsAfterSecondHalf2025Async();
         Task<List<Dictionary<string, object>>> ReadReservationsWithPaymentsAboveAsync(int minSum);
@@ -21,7 +21,7 @@ namespace DBSystemComparator_API.Repositories.Interfaces
         Task<int> UpdateClientsAddressAndPhoneAsync(bool isActive, DateTime dateThreshold);
         Task<int> UpdateRoomsPriceForReservationsAsync(int minCapacity, int priceIncrement);
         Task<int> UpdateServicesPriceAsync(int priceIncrement, bool isActive, int price);
-        Task<int> UpdatePriceForInactiveRoomsAsync(double discountMultiplier, int pricePerNight);
+        Task<int> UpdatePriceForInactiveRoomsAsync(double discountMultiplier, double pricePerNight);
         Task<int> UpdateRoomsPriceForReservationsToApril2024Async(int priceDecrement);
         // DELETE
         Task<int> DeletePaymentsOlderThanMarch2024Async();
@@ -31,7 +31,7 @@ namespace DBSystemComparator_API.Repositories.Interfaces
         Task<int> DeleteUnusedServicesPriceBelowAsync(int price);
         // HELPERS
         Task CreateClientsBatchAsync(IEnumerable<(string firstName, string secondName, string lastName, string email, DateTime dob, string address, string phone, bool isActive)> clients);
-        Task CreateRoomsBatchAsync(IEnumerable<(int number, int capacity, int pricePerNight, bool isActive)> rooms);
+        Task CreateRoomsBatchAsync(IEnumerable<(int number, int capacity, double pricePerNight, bool isActive)> rooms);
         Task CreateServicesBatchAsync(IEnumerable<(string name, int price, bool isActive)> services);
         Task CreateReservationsBatchAsync(IEnumerable<(int clientId, int roomId, DateTime checkIn, DateTime checkOut, DateTime creationDate)> reservations);
         Task CreatePaymentsBatchAsync(IEnumerable<(int reservationId, string description, int sum, DateTime creationDate)> payments);
